backend/requirements.txt

Django==5.0.6
django-ninja==1.1.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.22.0
openpyxl==3.1.3
uvicorn==0.30.1
python-multipart==0.0.9
